+++ modules/bucket-iam-policy/variables.tf
@@ -60,3 +60,9 @@ variable "arn_partition" {
   EOF
   default     = "aws"
 }
+
+variable "tags" {
+  type        = map(string)
+  description = "A map of tags to add to all resources."
+  default     = {}
+}